[PR Commit] Rename int-to-roman-num.py to int_to_roman_num.py
Rename int-to-roman-num.py to int_to_roman_num.py

This commit proposes a renaming of the file int-to-roman-num.py to int_to_roman_num.py in order to adhere to the naming conventions that respect the use of underscores (_) in Python file names. This change improves consistency and readability in the codebase.

Please review and merge this commit accordingly.

Thank you.